add config.py, rename ml_data.py -> main.py
